requirements.txt
streamlit>=1.35.0
playwright>=1.40.0
pandas>=2.0.0
langchain-groq>=0.1.0
langgraph>=0.2.0
langchain>=0.1.0
folium>=0.14.0
streamlit-folium>=0.15.0
requests>=2.31.0
pdfplumber>=0.10.0
pytesseract>=0.3.10
spacy>=3.7.0
pyttsx3>=2.90
Pillow>=10.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
